+++ tests/com/splunk/ServiceTest.java
@@ -22,9 +22,9 @@ import java.io.OutputStream;
 import java.net.Socket;
 import java.util.Arrays;
 import java.util.List;
-import org.junit.Test;
+import java.util.Map;
 
-import com.splunk.SDKTestCase.EventuallyTrueBehavior;
+import org.junit.Test;
 
 public class ServiceTest extends SDKTestCase {
     private static final String QUERY = "search index=_internal | head 10";
@@ -155,40 +155,97 @@ public class ServiceTest extends SDKTestCase {
         assertEquals(info.getService(), service);
     }
 
+    private void checkLoggedIn(Service service) {
+        ResponseMessage response;
+    	response = service.get("/services/authentication/users");
+        checkResponse(response);
+    }
+    
+    protected void checkNotLoggedIn(Service service) {
+    	try {
+            service.get("/services/authentication/users");
+            fail("Expected HttpException");
+        }
+        catch (HttpException e) {
+            assertEquals(401, e.getStatus());
+        }
+    }
+    
     @Test
     public void testLogin() {
-        ResponseMessage response;
-
         Service service = new Service(
                 (String) command.opts.get("host"),
                 (Integer) command.opts.get("port"),
                 (String) command.opts.get("scheme"));
 
         // Not logged in, should fail with 401
-        try {
-            response = service.get("/services/authentication/users");
-            fail("Expected HttpException");
-        }
-        catch (HttpException e) {
-            assertEquals(401, e.getStatus());
-        }
+        checkNotLoggedIn(service);
 
         // Logged in, request should succeed
         service.login(
                 (String) command.opts.get("username"),
                 (String) command.opts.get("password"));
-        response = service.get("/services/authentication/users");
-        checkResponse(response);
+        checkLoggedIn(service);
 
         // Logout, the request should fail with a 401
         service.logout();
-        try {
-            response = service.get("/services/authentication/users");
-            fail("Expected HttpException");
-        }
-        catch (HttpException e) {
-            assertEquals(401, e.getStatus());
-        }
+        checkNotLoggedIn(service);
+    }
+    
+    @Test
+    public void testLoginWithoutArguments() {
+    	ServiceArgs args = new ServiceArgs();
+    	args.setHost((String) command.opts.get("host"));
+    	args.setPort((Integer) command.opts.get("port"));
+    	args.setScheme((String) command.opts.get("scheme"));
+    	args.setUsername((String) command.opts.get("username"));
+    	args.setPassword((String) command.opts.get("password"));
+    	Service service = new Service(args);
+    	
+    	checkNotLoggedIn(service);
+    	
+    	service.login();
+    	checkLoggedIn(service);
+    	
+    	service.logout();
+    	checkNotLoggedIn(service);
+    }
+    
+    @Test
+    public void testLoginWithArgumentsOverridesServiceArgs() {
+    	ServiceArgs args = new ServiceArgs();
+    	args.setHost((String) command.opts.get("host"));
+    	args.setPort((Integer) command.opts.get("port"));
+    	args.setScheme((String) command.opts.get("scheme"));
+    	args.setUsername("I can't possibly be a user");
+    	args.setPassword("This password is nonsense.");
+    	Service service = new Service(args);
+    	
+    	checkNotLoggedIn(service);
+    	
+    	service.login(
+    		(String) command.opts.get("username"),
+    		(String) command.opts.get("password")
+    	);
+    	checkLoggedIn(service);
+    	
+    	service.logout();
+    	checkNotLoggedIn(service);
+    }
+    
+    @Test
+    public void testLoginWithoutAnyUsernameFails() {
+    	ServiceArgs args = new ServiceArgs();
+    	args.setHost((String) command.opts.get("host"));
+    	args.setPort((Integer) command.opts.get("port"));
+    	args.setScheme((String) command.opts.get("scheme"));
+    	Service service = new Service(args);
+    	
+    	try {
+    		service.login();
+    		fail("Expected IllegalStateException");
+    	}
+    	catch (IllegalStateException e) {}
     }
     
     @Test
@@ -533,47 +590,58 @@ public class ServiceTest extends SDKTestCase {
     public void testSearch() throws IOException {
         service.search(QUERY);    // throws no exception
 
-        InputStream jobOutput = service.search(QUERY,
-                new Args(),
-                new Args("output_mode", "json"));
+        Job job = service.search(QUERY, new Args());
+        while (!job.isDone()) {
+        	sleep(50);
+        }
+        
+        InputStream jobOutput = job.getResults();
         try {
-            char firstJsonChar;
-            if (service.versionCompare("5.0") < 0) {
-                firstJsonChar = '[';
-            } else {
-                firstJsonChar = '{';
-            }
-            
-            // Looks like JSON?
-            assertEquals(firstJsonChar, (char)jobOutput.read());
+        	ResultsReaderXml resultsReader = new ResultsReaderXml(jobOutput);
+        
+        	Map<String, String> event;
+        	int nEvents = 0;
+        
+        	do {
+        		event = resultsReader.getNextEvent();
+        		if (event != null) {
+        			nEvents += 1;
+        		}
+        	} while (event != null);
+        
+        	assertEquals(10, nEvents);
         }
         finally {
-            jobOutput.close();
+        	jobOutput.close();
         }
     }
     
     @Test
     public void testOneshot() throws IOException {
-        service.oneshot(QUERY); // throws no exception
+        service.oneshotSearch(QUERY); // throws no exception
         
-        InputStream jobOutput = service.oneshot(QUERY,
-                new Args("output_mode", "json"));
+        InputStream jobOutput = service.oneshotSearch(
+        	QUERY,
+            new Args("output_mode", "json")
+        );
+ 
         try {
-            char firstJsonChar;
-            if (service.versionCompare("5.0") < 0) {
-                firstJsonChar = '[';
-            } else {
-                firstJsonChar = '{';
-            }
-            
-            // Looks like JSON?
-            // NOTE: Not sure why a oneshot search would insert a leading
-            //       newline, whereas a blocking search doesn't.
-            assertEquals('\n', (char)jobOutput.read());
-            assertEquals(firstJsonChar, (char)jobOutput.read());
+        	ResultsReaderJson resultsReader = new ResultsReaderJson(jobOutput);
+        
+        	Map<String, String> event;
+        	int nEvents = 0;
+        
+        	do {
+        		event = resultsReader.getNextEvent();
+        		if (event != null) {
+        			nEvents += 1;
+        		}
+        	} while (event != null);
+        
+        	assertEquals(10, nEvents);
         }
         finally {
-            jobOutput.close();
+        	jobOutput.close();
         }
     }